requirements.txt
pandas
numpy
scikit-learn
xgboost
flask
flask-sqlalchemy
psycopg2-binary
kafka-python
requests
spacy
together
gunicorn
plotly
reportlab